Add files via upload
A volunteer put all the secret society members into an .xlsx sheet, but it needs to be formatted differently so developers can use it. This is the start of that project. Version 1 contains all the Skull and Bones members that could be found in the Utonian archives from 1911 through 1970 (which is when they stopped publishing the names for some reason). Updates will be coming soon, as the data is complete, but just needs to be formatted.